# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ambatuwork API v1",
        "description": "API documentation for Ambatuwork, a Scrum-based collaboration application used by Web and Android clients. Authenticate your requests using a Sanctum Bearer token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 Endpoint Base"
        }
    ],
    "paths": {
...
        "/projects/{project}/sprints/{sprint}/peer-review-cycle": {
            "get": {
                "tags": [
                    "Peer Reviews"
                ],
                "summary": "Get Peer Review Cycle details",
                "description": "Returns the peer review cycle details for the specified sprint.",
                "operationId": "0cc1e9d2c714b7f11cdfcde1575caf00",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PeerReviewCycle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project, Sprint or Cycle not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Peer Reviews"
                ],
                "summary": "Create/Open Peer Review Cycle",
                "description": "Creates and opens a peer review cycle for a sprint.",
                "operationId": "de30e0abd29a01a02615511a8eb9ba50",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "Sprint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cycle created/opened successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PeerReviewCycle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Sprint not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/peer-review-cycles/{cycle}/reviews": {
            "post": {
                "tags": [
                    "Peer Reviews"
                ],
                "summary": "Submit Peer Review",
                "description": "Submits a peer review for another user in the open cycle.",
                "operationId": "7293dbe50965996139c47947b7b27deb",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cycle",
                        "in": "path",
                        "description": "Peer Review Cycle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reviewee_user_id",
                                    "collaboration_score",
                                    "delivery_score",
                                    "communication_score"
                                ],
                                "properties": {
                                    "reviewee_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "collaboration_score": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "delivery_score": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "communication_score": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "continue_feedback": {
                                        "type": "string",
                                        "example": "Good planning.",
                                        "nullable": true
                                    },
                                    "improve_feedback": {
                                        "type": "string",
                                        "example": "None.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peer review submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PeerReview"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Cycle not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}/peer-review-cycles/{cycle}/close": {
            "post": {
                "tags": [
                    "Peer Reviews"
                ],
                "summary": "Close Peer Review Cycle",
                "description": "Closes an open peer review cycle.",
                "operationId": "c2877af9c1d10ebe73c7f11a56194db1",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cycle",
                        "in": "path",
                        "description": "Cycle ID to close",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cycle closed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PeerReviewCycle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Project or Cycle not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PeerReviewCycle": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sprint_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "example": "open"
                    },
                    "created_by_user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "PeerReview": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "peer_review_cycle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reviewer_user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "reviewee_user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "collaboration_score": {
                        "type": "integer",
                        "example": 5
                    },
                    "delivery_score": {
                        "type": "integer",
                        "example": 4
                    },
                    "communication_score": {
                        "type": "integer",
                        "example": 5
                    },
                    "continue_feedback": {
                        "type": "string",
                        "example": "Good planning.",
                        "nullable": true
                    },
                    "improve_feedback": {
                        "type": "string",
                        "example": "None.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token. Example: Bearer 1|abcdef...",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}